% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Logistic Regression Analysis}
\usage{
analysis(variables, adsl)
}
\arguments{
\item{variables}{ATRT and certian confounders (among SEX, B_WEIGHT, and AGE)
used in analysis}

\item{adsl}{the dataset.}
}
\value{
Regression analytical statistics.
}
\description{
Logistic Regression Model for ECOG Performance Status with
Treatment after adjusting for confounders
}
