% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_age_sex.R
\name{summary_age_sex}
\alias{summary_age_sex}
\title{Violin Plot Exploring Statistics of Age and Sex}
\usage{
summary_age_sex(adsl)
}
\arguments{
\item{adsl}{the dataset.}
}
\value{
Violin plot including summary stats.
}
\description{
Plot a violin plot showing summary satistics of age value
grouped by sex.
}
