% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ecog_trt.R
\name{summary_ecog_trt}
\alias{summary_ecog_trt}
\title{Histogram Exploring Statistics of ECOG Performance and Treatment}
\usage{
summary_ecog_trt(adsl)
}
\arguments{
\item{adsl}{the dataset.}
}
\value{
Histogram including summary stats.
}
\description{
Plot a histogram showing summary satistics of performance
status grouped by treatment.
}
