% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tbl.R
\name{summary_tbl}
\alias{summary_tbl}
\title{Summary Table for all required variables}
\usage{
summary_tbl(adsl)
}
\arguments{
\item{adsl}{the dataset.}
}
\value{
A summary table.
}
\description{
A table including frequency table for categorical
variables and summary for continuous variables.
}
