% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_weight_sex.R
\name{summary_weight_sex}
\alias{summary_weight_sex}
\title{Violin Plot Exploring Statistics of Weight and Sex}
\usage{
summary_weight_sex(adsl)
}
\arguments{
\item{adsl}{the dataset.}
}
\value{
Violin plot including summary stats.
}
\description{
Plot a violin plot showing summary satistics of weight value
grouped by sex.
}
