% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_weight_trt.R
\name{summary_weight_trt}
\alias{summary_weight_trt}
\title{Violin Plot Exploring Statistics of Weight and Treatment}
\usage{
summary_weight_trt(adsl)
}
\arguments{
\item{adsl}{the dataset.}
}
\value{
Violin plot including summary stats.
}
\description{
Plot a violin plot showing summary satistics of age value
grouped by treatment.
}
